requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
Faker==6.5.2
isort==5.7.0
lazy-object-proxy==1.5.2
Markdown==3.3.4
mccabe==0.6.1
mixer==7.1.2
psycopg2-binary==2.8.6
pylint==2.7.2
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
